Fixed an compile error

+++ src/main/java/battletank/controllers/ProjectileController.java
@@ -7,14 +7,14 @@ public class ProjectileController extends ActionController {
     //TODO: Add functionality to remove project if projectile healthpoints reaches 0.
     public ProjectileController(Projectile projectile) {
         //Assuming here that the plane hit is perfectly level (180 degrees).
-        int initialRotation = projectile.getRotation();
+        double initialRotation = projectile.getRotation();
         int totRotation = 180;
 
         //Calculate sigma, since (sigma - projectileRotation = 90 degrees)
-        int sigma = (totRotation / 2) - initialRotation;
+        double sigma = (totRotation / 2) - initialRotation;
 
         //Using the formula: initialRotation + 2*sigma + resultRotation = 180 degrees, we get:
-        int resRotation = initialRotation + (2 * sigma);
+        double resRotation = initialRotation + (2 * sigma);
 
         projectile.setRotation(resRotation);
